backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
django-filter
drf-spectacular
Pillow
factory_boy
